+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - VIMEET-2102 - Evaluate visio meeting
 - VIMEET-2128 - Show participant name on visio
+- VIMEET-2140 - Screen share without feedback
 
 ### Fixed
 - MV-236 - Add unavailability with a different timezone